Do lesson 3 activities and program 3.1

+++ src/ACT31.cpp
@@ -7,34 +7,32 @@ using namespace std;
 * Convert feet to inches
 * @param feet the number of feet
 * @return feet * 12
-*
 */
-int feetToInches(int feet) {
-	return (feet * 12);
+double feetToInches(double feet) {
+	return (feet * 12.0);
 }
 
 /**
 * Convert feet to centimeters
 * @param feet the number of feet
 * @return feet * 30.5
-*
 */
-double feetToCM(int feet) {
+double feetToCM(double feet) {
 	return (feet * 30.5);
 }
 
 int main() {
-	int feet;
+	double feet;
 	cout << "Please enter a number of feet: ";
 	cin >> feet;
 
-	int inches;
+	double inches;
 	inches = feetToInches(feet);
 
 	double centimeters;
 	centimeters = feetToCM(feet);
 
-	printf("\nThere are %d inches and %.1f centimeters in %d feet.",
+	printf("\nThere are %.10g inches and %.10g centimeters in %.10g feet.",
 			inches, centimeters, feet);
 
 	return 0;
